Extract insertNoderelativeToElement
