+++ lib/index.js
@@ -1,6 +1,9 @@
 import getPosts from './getPosts.js';
 import groupBy from './groupBy.js';
+import { ElementInsertionPosition, insertNodeRelativeToElement } from './insertNodeRelativeToElement.js';
 import times from './times.js';
+
+export { ElementInsertionPosition } from './insertNodeRelativeToElement.js';
 
 
 /**
@@ -39,19 +42,6 @@ import times from './times.js';
  * @property {!string} selector A CSS selector string for an element to insert relative to.
  * @property {!ElementInsertionPosition} position The position to insert a node relative to an element.
  */
-
-
-/**
- * The position for insertion of an element.
- * @readonly
- * @enum {string}
- */
-export const ElementInsertionPosition = {
-    BEFORE: 'before',
-    AFTER: 'after',
-    BEGIN: 'begin',
-    END: 'end'
-};
 
 
 /**
@@ -104,29 +94,6 @@ export function SeriesDisplay(ghostContentApi) {
                 )
             )
         );
-    }
-
-    function insertNodeRelativeToElement(node, element, insertionPosition) {
-        switch (insertionPosition) {
-            case ElementInsertionPosition.BEFORE:
-                element.before(node);
-                break;
-
-            case ElementInsertionPosition.AFTER:
-                element.after(node);
-                break;
-
-            case ElementInsertionPosition.BEGIN:
-                element.prepend(node);
-                break;
-
-            case ElementInsertionPosition.END:
-                element.append(node);
-                break;
-
-            default:
-                throw new RangeError('insertionPosition must be one of ElementInsertionPosition.');
-        }
     }
 
     function normailizeBuildOptions(options, defaultSeriesTagSlugFinder) {
